+++ src/test/groovy/ru/vyarus/java/generics/resolver/cases/methodgeneric/support/Err.groovy
@@ -1,0 +1,12 @@
+package ru.vyarus.java.generics.resolver.cases.methodgeneric.support
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 07.05.2018
+ */
+class Err {
+
+    void errMeth() {
+
+    }
+}