+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/DeclarationType.groovy
@@ -1,0 +1,18 @@
+package ru.vyarus.java.generics.resolver.inlying.support
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 07.05.2018
+ */
+// use type generics to check transitive resolution
+class DeclarationType<Fld, Ret, Par> {
+
+    private SubType<Fld> one
+
+    public SubType<Ret> ret() {
+        return null;
+    }
+
+    public void param(SubType<Par> par) {
+    }
+}