+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/Err.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.java.generics.resolver.inlying.support
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 07.05.2018
+ */
+class Err {
+    String wrongField
+}